Fix the import path and move files to form folder.
